More recipe mechanics!
Implemented: Shapeless recipes, some more crafttweaker stuff, but
crafttweaker is still not ready. There's a new class acting as a base
for recipes - BaseCubicCraftingRecipe - which adds some basic stuff.
Ingredient has also moved there, and for debug, there's a new method.
Recipes for the totem of undieing and beacon have been added, vanilla
beacon recipe removed. More to come + config for all these.

Fixed: A bug with the orientation of the slot matrix

Still to come: Maybe a config (soon (tm)) for recipes, and actual
recipes for the blocks added by the mod. Also more cross mod interaction
with (including, but not limited to) Botania, Thermal series/EnderIO and
eventually ProjectE.